Start using Double instead of Decimal in Atom
Respectively, replace Double with Decimal in all places of the code it’s used. Tests showed that machine epsilon of Double is small enough to be as precise as Decimal in computations. In other words, Double has enough precision to not to break algorithm which relies tight on strict equality of two Double elements. Consider solving the issue (if one shows up) by creating the “approximately equal” operator (~=) as it was done in VectorMath lib: https://github.com/nicklockwood/VectorMath#operators

+++ ClusterAlgorithm/Algorithm/ClusterAlgorithm.swift
@@ -9,7 +9,7 @@ import Foundation
 
 enum ClusterMetric {
     case count(UInt)
-    case radius(Decimal)
+    case radius(Double)
 }
 
 class ClusterAlgorithm {
